Fixed minor errors.

+++ core/beam_decoder.py
@@ -50,7 +50,8 @@ class BeamSearchDecoder(object):
             done_scores_max, done_parent_indices = torch.max(end_scores, -1)
             done_symbols = torch.cat([torch.squeeze(torch.gather(beam_symbols, 1,
                                       done_parent_indices.view(-1, 1, 1).repeat(1, 1, t + 1)), 1),
-                                      torch.full([batch_size, self.n_time_steps - t], self._end, device=self.device)], -1)
+                                      torch.full([batch_size, self.n_time_steps - t], 
+                                            self._end, dtype=torch.int64, device=self.device)], -1)
 
             cand_mask = (done_scores_max >= k_scores[:, -1]) & (~cand_finished | (done_scores_max > cand_scores))
             cand_finished = cand_mask | cand_finished
